\docType{package}
\name{mchof}
\alias{mchof}
\alias{mchof-package}
\title{mchof: multicore higher-order-functions.}
\description{
	mchof (pronounced mac'hoff) provides easy-to-use, consistent parallel implementations of common higher-order functions. Map-like functions - such as those found in plyr - are widely known and loved by useR's. This package aims to bring parallel, predictable and simple implementations of higher order functions found in other languages to R.

}
