\name{mcUnion}
\alias{mcIntersect}
\alias{mcIsElement}
\alias{mcSetDiff}
\alias{mcSetEqual}
\alias{mcSymDiff}
\alias{mcUnion}
\title{Higher-Order-Functions for Set Theory}
\usage{
  mcUnion(f, g)

  mcIntersect(f, g)

  mcSetDiff(f, g)

  mcSymDiff(f, g)

  mcSetEqual(f, g)

  mcIsElement(f, g)

  mcSymDiff(f, g)
}
\arguments{
  \item{f}{a function, or a string giving the name of a
  function.}

  \item{g}{a function, or a string giving the name of a
  function.}
}
\description{
  1
}
\details{
  1
}
\examples{

NULL
}

