\name{mcIndMap}
\alias{mcIndMap}
\title{mcIndMap}
\usage{
  mcIndMap(f, x, paropts = NULL)
}
\arguments{
  \item{f}{a binary function that takes an index as its
  first parameters and a value as its second parameter, or
  a string giving the name of such a function.}

  \item{x}{a vector or list which will be used as an
  initial input for f.}

  \item{paropts}{}
}
\description{
  mcIndMap
}
\keyword{mcIndMap}

