\name{mcUnfold}
\alias{mcUnfold}
\title{mcUnfold}
\usage{
  mcUnfold(f, x, paropts = NULL)
}
\arguments{
  \item{f}{a function that generates a list}

  \item{first}{an initial value for the first position of
  f}

  \item{paropts}{paropts a list of parameters to be handed
  to mclapply (see details and \code{\link{mclapply}})}
}
\description{
  mcUnfold
}
\details{
  details?
}

