\docType{package}
\name{plyr}
\alias{plyr}
\alias{plyr-package}
\title{mchof: multicore higher-order-functions}
\description{
  mchof: multicore higher-order-functions
}
\section{Functions}{
  Below is a full list of functions currently included in
  mchof, along with a description of what they do

  \itemize{ \item \code{\link{mcFilter}}: get the elements
  of a list matching a predicate. \item
  \code{\link{mcReduce}}: applies an associative binary
  operator to a list (eg. +, *, rbind), returning a single
  element. \item \code{\link{mcFind}}: get the first
  element in a list matching a predicate. \item
  \code{\link{mcPosition}}: find the first element in a
  list matching a predicate. \item \code{\link{mcZipWith}}:
  combine n lists into a list of n-tuples, apply a function
  to each element of the result. \item \code{\link{mcZip}}:
  combine n lists into a list of n-tuples. \item
  \code{\link{mcUnzipWith}}: split a list of n-tuples into
  n lists, apply a function to each element of the results.
  \item \code{\link{mcUnzip}}: split a list of n-tuples
  into n lists. }
}

