\name{mcIndMap}
\alias{mcIndMap}
\title{Map-like Higher-Order-Functions}
\usage{
  mcIndMap(f, x, paropts = NULL)
}
\arguments{
  \item{p}{1}

  \item{f}{1}

  \item{x}{x}

  \item{paropts}{1}
}
\description{
  Map-like Higher-Order-Functions
}
\details{
  1
}
\examples{

NULL
}
\keyword{mcIndMap}
\keyword{mcIterateWhile,}

